[SCRUM-172] refactor: interceptor 위치 변경
