+++ src/main/java/com/projectlyrics/server/domain/auth/authentication/interceptor/SlackInterceptor.java
@@ -1,4 +1,4 @@
-package com.projectlyrics.server.domain.auth.authentication;
+package com.projectlyrics.server.domain.auth.authentication.interceptor;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
